refactor(dashboard): split fsm-hub-diagnostics into lane + invariant analysis
The 422-line diagnostics module combined two distinct concerns: per-lane
state interpretation (stall detection, tone mapping, summaries) and
cross-cutting operational insight (invariant checks, next-step hints,
recovery analysis). Split along that boundary.

- fsm-hub-lane-analysis.ts (179): isObservedStall, laneMeaning,
  deriveObservedLaneSummaries — per-lane classification
- fsm-hub-invariant-analysis.ts (249): brokenInvariantKey, invariantDetail,
  nextExpectedStep, deriveOperationalInsight, invariantRows — invariant
  and insight decisions (imports lane analysis for stalled-lane evidence)

Consumers updated: fsm-hub.ts (re-exports), fsm-hub-pipeline-panels.ts
(deriveOperationalInsight, deriveObservedLaneSummaries), fsm-hub-health-panels.ts
(invariantRows). No behavior change. 95/95 test files, 655/655 tests pass.
